Вынесен экран настроек тренировки из MoreScreen
- Создан WorkoutSettingsScreen с настройками отдыха и уведомлений
- Заменён DisclosureGroup на NavigationLink в MoreScreen
- Добавлен кейс workoutSettings в аналитику
- Обновлены UI- и unit-тесты
