+++ SwiftUI-SotkaApp/Screens/More/MoreScreen.swift
@@ -1,4 +1,3 @@
-import OSLog
 import SwiftData
 import SwiftUI
 import SWUtils
@@ -11,8 +10,6 @@ struct MoreScreen: View {
     @Environment(\.analyticsService) private var analytics
     @Environment(AuthHelperImp.self) private var authHelper
     @Environment(\.isReadOnlyMode) private var isReadOnlyMode
-    @AppStorage(Key.isWorkoutGroupExpanded.rawValue) private var isWorkoutGroupExpanded = true
-    @AppStorage(Key.isWorkoutRestGroupExpanded.rawValue) private var isWorkoutRestGroupExpanded = true
     let user: User
     @State private var aboutInfopost: Infopost?
     @State private var showResetDialog = false
@@ -36,7 +33,7 @@ struct MoreScreen: View {
                     #if DEBUG
                     debugCurrentDayPicker
                     #endif
-                    workoutSettingsGroup
+                    workoutSettingsButton
                     if !isOfflineUser {
                         syncJournalButton
                     }
@@ -62,10 +59,6 @@ struct MoreScreen: View {
                     githubButton
                 }
             }
-            .animation(.default, value: appSettings.workoutNotificationsEnabled)
-            .animation(.default, value: appSettings.playTimerSound)
-            .animation(.default, value: isWorkoutGroupExpanded)
-            .animation(.default, value: isWorkoutRestGroupExpanded)
             .navigationTitle(.more)
             .navigationBarTitleDisplayMode(.inline)
             .trackScreen(.more)
@@ -73,12 +66,6 @@ struct MoreScreen: View {
                 if aboutInfopost == nil {
                     aboutInfopost = statusManager.infopostsService.loadAboutInfopost()
                 }
-            }
-            .onChange(of: appSettings.workoutNotificationsEnabled) { _, _ in
-                analytics.log(.userAction(action: .toggleWorkoutNotifications))
-            }
-            .onChange(of: appSettings.restTime) { _, newValue in
-                analytics.log(.userAction(action: .selectRestTime(seconds: newValue)))
             }
         }
     }
@@ -155,81 +142,11 @@ private extension MoreScreen {
     }
     #endif
 
-    var workoutSettingsGroup: some View {
-        DisclosureGroup(isExpanded: $isWorkoutGroupExpanded) {
-            @Bindable var settings = appSettings
-            NavigationLink(destination: CustomExercisesScreen()) {
-                Text(.customExercises)
-            }
-            .accessibilityIdentifier("customExercisesButton")
-            notificationToggle
-            if settings.workoutNotificationsEnabled {
-                makeNotificationTimePicker(
-                    $settings.workoutNotificationTime
-                )
-            }
-            DisclosureGroup(.moreScreenRestGroup, isExpanded: $isWorkoutRestGroupExpanded) {
-                makeRestTimePicker($settings.restTime)
-                makeTimerSoundToggle($settings.playTimerSound)
-                if settings.playTimerSound {
-                    makeTimerSoundPicker($settings.timerSound)
-                }
-                if UIDevice.current.userInterfaceIdiom == .phone {
-                    makeVibrateToggle($settings.vibrate)
-                }
-            }
-        } label: {
-            Text(.moreScreenWorkoutGroup)
-                .accessibilityIdentifier("moreScreenWorkoutGroup")
+    var workoutSettingsButton: some View {
+        NavigationLink(destination: WorkoutSettingsScreen()) {
+            Text(.workoutSettingsScreen)
         }
-    }
-
-    @ViewBuilder
-    var notificationToggle: some View {
-        @Bindable var settings = appSettings
-        Toggle(
-            .workoutNotifications,
-            isOn: .init(
-                get: { appSettings.workoutNotificationsEnabled },
-                set: {
-                    appSettings.setWorkoutNotificationsEnabled($0)
-                }
-            )
-        )
-    }
-
-    func makeNotificationTimePicker(_ value: Binding<Date>) -> some View {
-        DatePicker(
-            .notificationTime,
-            selection: value,
-            displayedComponents: .hourAndMinute
-        )
-    }
-
-    func makeTimerSoundToggle(_ value: Binding<Bool>) -> some View {
-        Toggle(.timerSoundToggle, isOn: value)
-    }
-
-    func makeVibrateToggle(_ value: Binding<Bool>) -> some View {
-        Toggle(.timerVibrateToggle, isOn: value)
-    }
-
-    func makeTimerSoundPicker(_ value: Binding<TimerSound>) -> some View {
-        Picker(.moreScreenTimerSound, selection: value) {
-            ForEach(TimerSound.allCases, id: \.self) { sound in
-                Text(sound.displayName).tag(sound)
-            }
-        }
-        .pickerStyle(.navigationLink)
-    }
-
-    func makeRestTimePicker(_ value: Binding<Int>) -> some View {
-        Picker(.restTimePicker, selection: value) {
-            ForEach(Constants.restPickerOptions, id: \.self) { seconds in
-                Text(RestTimeComponents(totalSeconds: seconds).localizedString).tag(seconds)
-            }
-        }
-        .pickerStyle(.navigationLink)
+        .accessibilityIdentifier("workoutSettingsButton")
     }
 
     var feedbackButton: some View {
@@ -335,13 +252,6 @@ private extension MoreScreen {
             Text(.moreScreenSyncJournalButton)
         }
         .accessibilityIdentifier("syncJournalButton")
-    }
-}
-
-private extension MoreScreen {
-    enum Key: String {
-        case isWorkoutGroupExpanded = "WorkoutSettings.Expanded"
-        case isWorkoutRestGroupExpanded = "WorkoutSettings.Rest.Expanded"
     }
 }
