+++ SwiftUI-SotkaApp/SupportingFiles/Localizable.xcstrings
@@ -2905,25 +2905,6 @@
         }
       }
     },
-    "MoreScreen.Workout.Group" : {
-      "comment" : "The title for a group of settings related to workouts.",
-      "extractionState" : "manual",
-      "isCommentAutoGenerated" : true,
-      "localizations" : {
-        "en" : {
-          "stringUnit" : {
-            "state" : "translated",
-            "value" : "Workout"
-          }
-        },
-        "ru" : {
-          "stringUnit" : {
-            "state" : "translated",
-            "value" : "Тренировки"
-          }
-        }
-      }
-    },
     "New password" : {
       "extractionState" : "manual",
       "localizations" : {
@@ -6021,6 +6002,25 @@
           "stringUnit" : {
             "state" : "translated",
             "value" : "Тренировка"
+          }
+        }
+      }
+    },
+    "WorkoutSettingsScreen" : {
+      "comment" : "The title for a group of settings related to workouts.",
+      "extractionState" : "manual",
+      "isCommentAutoGenerated" : true,
+      "localizations" : {
+        "en" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "Workout"
+          }
+        },
+        "ru" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "Тренировки"
           }
         }
       }
